% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cie11cl-package.R
\docType{package}
\name{cie11cl-package}
\alias{cie11cl}
\alias{cie11cl-package}
\title{cie11cl: ICD-11 (CIE-11) Access and ICD-10 to ICD-11 Crosswalks for Chile}
\description{
Deterministic access to the World Health Organization International Classification of Diseases, 11th Revision (ICD-11, MMS linearization) in Spanish, with code lookup, lexical search, hierarchical navigation, cluster validation, and ICD-10 to ICD-11 crosswalks with traceable certainty levels. The package bundles the WHO MMS 2026 linearization (37,212 entities). The same input plus the same versioned reference data always yields the same output.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/RodoTasso/cie11cl}
  \item Report bugs at \url{https://github.com/RodoTasso/cie11cl/issues}
}

}
\author{
\strong{Maintainer}: Rodolfo Tasso \email{rtasso@uchile.cl}

Authors:
\itemize{
  \item Rodolfo Tasso \email{rtasso@uchile.cl}
}

}
\keyword{internal}
